requirements.txt
# YOLOv11 花粉识别与计数项目依赖

# 核心依赖
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.23.0
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置文件
PyYAML>=6.0

# 其他工具
tqdm>=4.65.0
requests>=2.31.0
scikit-learn>=1.2.0
